# Unify text format and speed-up finding steps offset