--- a/include/chemfiles/formats/MOL2.hpp
+++ b/include/chemfiles/formats/MOL2.hpp
@@ -5,9 +5,7 @@
 #define CHEMFILES_FORMAT_MOL2_HPP
 
 #include <iosfwd>
-#include <memory>
 #include <string>
-#include <vector>
 #include <unordered_map>
 
 #include "chemfiles/File.hpp"
@@ -23,14 +21,14 @@ class Frame;
 /// `END` records, separating the steps.
 ///
 /// [MOL2]:
-class MOL2Format final: public Format {
+class MOL2Format final: public TextFormat {
 public:
-    MOL2Format(std::string path, File::Mode mode, File::Compression compression);
+    MOL2Format(std::string path, File::Mode mode, File::Compression compression):
+        TextFormat(std::move(path), mode, compression) {}
 
-    void read_step(size_t step, Frame& frame) override;
-    void read(Frame& frame) override;
-    void write(const Frame& frame) override;
-    size_t nsteps() override;
+    void read_next(Frame& frame) override;
+    void write_next(const Frame& frame) override;
+    std::streampos forward() override;
 
 private:
     // Read Atoms
@@ -39,12 +37,8 @@ private:
     // Read Bonds
     void read_bonds(Frame& frame, size_t nbonds);
 
-    std::unique_ptr<TextFile> file_;
     /// Map of residues, indexed by residue id.
     std::unordered_map<size_t, Residue> residues_;
-    /// Storing the positions of all the steps in the file, so that we can
-    /// just `seekg` them instead of reading the whole step.
-    std::vector<std::streampos> steps_positions_;
 };
 
 template<> FormatInfo format_information<MOL2Format>();