--- a/include/chemfiles/formats/PDB.hpp
+++ b/include/chemfiles/formats/PDB.hpp
@@ -6,7 +6,6 @@
 
 #include <map>
 #include <iosfwd>
-#include <memory>
 #include <string>
 #include <vector>
 #include <tuple>
@@ -24,16 +23,16 @@ class Frame;
 /// `END` records, separating the steps.
 ///
 /// [PDB]: ftp://ftp.wwpdb.org/pub/pdb/doc/format_descriptions/Format_v33_A4.pdf
-class PDBFormat final: public Format {
+class PDBFormat final: public TextFormat {
 public:
-    PDBFormat(std::string path, File::Mode mode, File::Compression compression);
+    PDBFormat(std::string path, File::Mode mode, File::Compression compression):
+        TextFormat(std::move(path), mode, compression) {}
 
     ~PDBFormat() override;
 
-    void read_step(size_t step, Frame& frame) override;
-    void read(Frame& frame) override;
-    void write(const Frame& frame) override;
-    size_t nsteps() override;
+    void read_next(Frame& frame) override;
+    void write_next(const Frame& frame) override;
+    std::streampos forward() override;
 
     // Connect residues based on a predefined table
     static void link_standard_residue_bonds(Frame& frame);
@@ -53,19 +52,14 @@ private:
     // Runs when a chain is terminated to update residue information
     void chain_ended(Frame& frame);
 
-    std::unique_ptr<TextFile> file_;
-
     /// This typedef represent the 'full' name of a residue, this means the
     /// chainid, the residue sequence id, and the residue insertion code
     using residue_info = std::tuple<char, size_t, char>;
 
     /// Map where the key is the residue_info and the value is a residue object
     std::map<residue_info, Residue> residues_;
-    /// Storing the positions of all the steps in the file, so that we can
-    /// just `seekg` them instead of reading the whole step.
-    std::vector<std::streampos> steps_positions_;
     /// Number of models written/read to the file.
-    size_t models_;
+    size_t models_ = 0;
     /// List of all atom offsets. This maybe pushed in read_ATOM or if a TER
     /// record is found. It is reset every time a frame is read.
     std::vector<size_t> atom_offsets_;