--- a/src/File.cpp
+++ b/src/File.cpp
@@ -41,8 +41,9 @@ TextFile::TextFile(std::string path, File::Mode mode, File::Compression compress
 /// line markers (`\n`, `\r`, `\r\n`).
 ///
 /// Code from http://stackoverflow.com/a/6089413
-void TextFile::get_line(std::string& string) {
+void TextFile::get_line_impl(std::string& string) {
     string.clear();
+    string.reserve(128);
 
     // The characters in the stream are read one-by-one using a std::streambuf.
     // That is faster than reading them one-by-one using the std::istream.
@@ -50,7 +51,7 @@ void TextFile::get_line(std::string& string) {
     // The sentry object performs various tasks,
     // such as thread synchronization and updating the stream state.
     std::istream::sentry sentry(*this, true);
-    std::streambuf* buffer = std::istream::rdbuf();
+    std::streambuf* buffer = this->rdbuf();
 
     while(true) {
         int c = buffer->sbumpc();
@@ -65,7 +66,7 @@ void TextFile::get_line(std::string& string) {
         case EOF:
             // Also handle the case when the last line has no line ending
             if(string.empty()) {
-                std::istream::setstate(std::ios::eofbit);
+                this->setstate(std::ios::eofbit);
             }
             return;
         default:
@@ -74,24 +75,45 @@ void TextFile::get_line(std::string& string) {
     }
 }
 
-void TextFile::rewind() {
-    std::istream::clear();
-    std::istream::seekg(0, std::ios::beg);
+void TextFile::skip_line_impl() {
+    std::istream::sentry sentry(*this, true);
+    std::streambuf* buffer = this->rdbuf();
+
+    while(true) {
+        int c = buffer->sbumpc();
+        switch (c) {
+        case '\n':
+            return;
+        case '\r':
+            if(buffer->sgetc() == '\n') {
+                buffer->sbumpc();
+            }
+            return;
+        case EOF:
+            this->setstate(std::ios::eofbit);
+            return;
+        default:
+            continue;
+        }
+    }
 }
 
-bool TextFile::eof() {
-    return std::istream::eof();
+void TextFile::rewind() {
+    this->clear();
+    this->seekg(0, std::ios::beg);
 }
 
 std::string TextFile::readline() {
     // Disable exceptions checking, and manually check bellow
     auto state = this->exceptions();
     this->exceptions(std::fstream::goodbit);
+
     std::string line;
-    get_line(line);
+    get_line_impl(line);
     if (this->fail()) {
         throw file_error("could not read a line in {}", this->path());
     }
+
     // Re-enable exceptions checking
     this->exceptions(state);
     return line;
@@ -104,13 +126,46 @@ std::vector<std::string> TextFile::readlines(size_t n) {
 
     auto lines = std::vector<std::string>(n);
     for (size_t i = 0; i < n; i++) {
-        get_line(lines[i]);
+        get_line_impl(lines[i]);
     }
 
     if (this->fail()) {
         throw file_error("could not read a line in {}", this->path());
     }
+
     // Re-enable exceptions checking
     this->exceptions(state);
+
     return lines;
+}
+
+void TextFile::skipline() {
+    // Disable exceptions checking, and manually check bellow
+    auto state = this->exceptions();
+    this->exceptions(std::fstream::goodbit);
+
+    skip_line_impl();
+    if (this->fail()) {
+        throw file_error("could not read a line in {}", this->path());
+    }
+
+    // Re-enable exceptions checking
+    this->exceptions(state);
+}
+
+void TextFile::skiplines(size_t n) {
+    // Disable exceptions checking, and manually check bellow
+    auto state = this->exceptions();
+    this->exceptions(std::fstream::goodbit);
+
+    for (size_t i = 0; i < n; i++) {
+        skip_line_impl();
+    }
+
+    if (this->fail()) {
+        throw file_error("could not read a line in {}", this->path());
+    }
+
+    // Re-enable exceptions checking
+    this->exceptions(state);
 }