… src/formats/CSSR.cpp
… src/formats/GRO.cpp
… src/formats/LAMMPSData.cpp
… src/formats/MOL2.cpp
… src/formats/SDF.cpp
… src/formats/SMI.cpp
… src/formats/XYZ.cpp
… tests/CMakeLists.txt
… tests/formats/amber-netcdf.cpp
… tests/formats/cml.cpp
… tests/formats/cssr.cpp
… tests/formats/dcd-molfile.cpp
… tests/formats/gro.cpp
… tests/formats/lammps-data.cpp
… tests/formats/lammps-molfile.cpp
… tests/formats/mmcif.cpp
… tests/formats/mmtf.cpp
… tests/formats/mol2.cpp
… tests/formats/pdb.cpp
… tests/formats/sdf.cpp
… tests/formats/smi.cpp
… tests/formats/tinker.cpp
… tests/formats/tng.cpp
… tests/formats/trr-molfile.cpp
… tests/formats/xtc-molfile.cpp
… tests/formats/xyz.cpp
… tests/trajectory.cpp