The first part of this PR unify all the text formats (except mmCIF, because the logic to find offsets is a bit complex) under a new abstract class `TextFormat`. This new class is responsible for finding the "start of step" offsets in the file.


The second part add new functions to TextFile to just skip a (few) lines. This is useful when looking for offset, as we don't have to allocate memory for string we immediately throw away.

---

Using the new code, I get up to 3x speedup with a C++ benchmark.

This PR was motivated by https://github.com/MDAnalysis/mdanalysis/pull/1862#issuecomment-509199512, as it looks like chemfiles is slower than MDAnalysis when reading XYZ.

With this PR, I get nice gains with a small file containing a lot of frames:
```python
In [4]: mda.coordinates.XYZ.XYZReader('./water.xyz')
Out[4]: <XYZReader ./water.xyz with 100 frames of 297 atoms>

In [5]: %timeit mda.coordinates.XYZ.XYZReader('./water.xyz')
1.08 ms ± 48.3 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

# Before
In [8]: %timeit chemfiles.ChemfilesReader('./water.xyz')
15.4 ms ± 215 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

# After
In [2]: %timeit chemfiles.ChemfilesReader('./water.xyz')
5.09 ms ± 84.8 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

But not so much for bigger files with a single frame.
```python
In [3]: mda.coordinates.XYZ.XYZReader('./big.xyz')
Out[3]: <XYZReader ./big.xyz with 1 frames of 29997 atoms>

In [4]: %timeit mda.coordinates.XYZ.XYZReader('./big.xyz')
55.3 ms ± 1.79 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

# Before
In [3]: %timeit chemfiles.ChemfilesReader('./big.xyz', format='XYZ')
87.2 ms ± 999 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

# After
In [5]: %timeit chemfiles.ChemfilesReader('./big.xyz', format='XYZ')
74.6 ms ± 680 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

We are still slower than MDAnalysis, but there is also quite a lot of overhead when doing C++ => C => Python (chemfiles) => Python (MDAnalysis).

Overall, this PR is still a win and I think worth to merge.