@Luthaf 2019-07-08
@richardjgowers, this might interest you =) the improvement should be in the next version (0.9.2)