@Luthaf 2019-07-09
So I can confirm that locally, a C++ program using chemfiles is as fast as MDAnalysis when reading an XYZ file. 

Using chemfiles from MDAnalysis is still slower, but the difference must come from the conversion from C++ to C to Python.