@Luthaf 2019-07-09
>  have you tried a pure C++ benchmark? That should be faster than Python!

Yes, the pure C++ benchmark is as fast as Python, not faster. This makes sense, as all Python is doing here is calling into the standard library (written in C) to read lines and parse floats. So I would not expect high gains by simply using c++, even more so that this code is IO bound, not really CPU bound.

> You mentioned that the C++ benchmark is 3x faster, but I just want to make sure!

That's 3x faster than current master =)