@frodofine 2019-07-09
Ok, that makes sense. XYZ files are quite simple, so there's probably little to gain from a C++ vs Python.